defined the tomoLoResFactor, included a script to analyse all folders in a path

+++ automation/qualityMeasures.py
@@ -14,7 +14,7 @@ import xraySimulation
 import filterPerformance
 import beamHardeningSimulation as bhs
 import scatteringSimulation as ss
-import resolutionEstimation as re
+import resolutionEstimation as resEst
 import snrTest
 
 
@@ -48,25 +48,28 @@ class QualityMeasuresAnalyzer:
     - computeResolution: Resolution based on decorrelation analysis
     - analyseAll: Perform all analyses and save results to a file
     """
-    def __init__(self, directoryPath, sampleMaterial, shape):
+    def __init__(self, directoryPath: str, sampleMaterial: str, shape: tuple[int, int]):
         self.directoryPath = directoryPath
         self.sampleMaterial = sampleMaterial
         self.shape = shape
-        (self.sampleDiameterMm, self.filterThicknessMm, self.vxSizeMm, self.kvp, self.dfAverage,
+        (self.sampleDiameterMm, self.filterThicknessMm, self.vxSizeMm, self.kvp, self.binning, self.dfAverage,
          self.cfAverage, self.kfMiddle, self.tomoLoRes) = \
             extractData.extractAllData(self.directoryPath, shape=self.shape, offset=10000)
         self.tomoSliceZ = self.tomoLoRes[self.tomoLoRes.shape[1] // 2]
         self.tomoSliceX = self.tomoLoRes[:, self.tomoLoRes.shape[0] // 2]
         self.tomoSliceY = self.tomoLoRes[:, :, self.tomoLoRes.shape[0] // 2]
 
+        self.tomoLoResFactor = np.floor(self.kfMiddle.shape[0] / self.tomoSliceZ.shape[0])
+
     def computeBhc(self):
         """compute beam hardening correction coefficient"""
         sampleDiameterVx = self.sampleDiameterMm / self.vxSizeMm
-        bhcTheo, _, tomoSimu, _ = bhs.simulateBH(sampleDiameterMm=self.sampleDiameterMm, sampleDiameterVx=sampleDiameterVx, kvp=self.kvp,
+        bhcTheo, _, tomoSimu, _ = bhs.simulateBH(sampleDiameterMm=self.sampleDiameterMm,
+                                                 sampleDiameterVx=sampleDiameterVx, kvp=self.kvp,
                                                  filterMaterial='Fe', filterThicknessMm=self.filterThicknessMm,
                                                  materialName=self.sampleMaterial, plotCurve=False, verbose=False)
         bhcSimu = bhs.estimateBhcFromTomo(tomoSimu, vxSzMm=self.vxSizeMm, plot=False, verbose=False)
-        bhc = bhs.estimateBhcFromTomo(self.tomoSliceZ, vxSzMm=2 * self.vxSizeMm, plot=False, verbose=False)
+        bhc = bhs.estimateBhcFromTomo(self.tomoSliceZ, vxSzMm=self.tomoLoResFactor * self.vxSizeMm, plot=False, verbose=False)
 
         print("\n#### BHC Result ####")
         print("Experimental BHC = %.4f" % bhc)
@@ -79,7 +82,7 @@ class QualityMeasuresAnalyzer:
         """compute the scattering contribution"""
         transExpe, transTheo, estimatedSampleDiameterMm, result = \
             ss.calcScatteringFromData(self.kfMiddle, self.cfAverage, self.dfAverage, self.tomoSliceZ, self.kvp, 'Fe',
-                                      self.filterThicknessMm, self.sampleMaterial, 2 * self.vxSizeMm, offset=0)
+                                      self.filterThicknessMm, self.sampleMaterial, self.tomoLoResFactor * self.vxSizeMm, offset=0)
 
         energyKeV, spectrum = filterPerformance.setSpectrum(self.kvp, 'Fe', self.filterThicknessMm)
         scatEsti = ss.estimateMeasuredScatterAsPercentOfFlux(energyKeV, spectrum, self.sampleMaterial, self.sampleDiameterMm)
@@ -112,9 +115,9 @@ class QualityMeasuresAnalyzer:
 
     def computeResolution(self):
         """compute the resolution of the tomographic dataset"""
-        resX = re.findImageRes(self.tomoSliceX, pxSzMm=2 * self.vxSizeMm, Ng=8, plot=False)
-        resY = re.findImageRes(self.tomoSliceY, pxSzMm=2 * self.vxSizeMm, Ng=8, plot=False)
-        resZ = re.findImageRes(self.tomoSliceZ, pxSzMm=2 * self.vxSizeMm, Ng=8, plot=False)
+        resX = resEst.findImageRes(self.tomoSliceX, pxSzMm=self.tomoLoResFactor * self.vxSizeMm, Ng=10, plot=False)
+        resY = resEst.findImageRes(self.tomoSliceY, pxSzMm=self.tomoLoResFactor * self.vxSizeMm, Ng=10, plot=False)
+        resZ = resEst.findImageRes(self.tomoSliceZ, pxSzMm=self.tomoLoResFactor * self.vxSizeMm, Ng=10, plot=False)
         # resZ is much lower than the other two, similar to the SNR result without masking
 
         print("\n#### Resolution Result ####")
@@ -124,18 +127,32 @@ class QualityMeasuresAnalyzer:
 
         return [resX, resY, resZ]
 
+
     def analyseAll(self):
         """
-        Perform the analysis of the quality measures and save both the results and the log to the parent directory
+        Perform the analysis of the quality measures and save both the results and the log to the parent/results directory
         :return: bhc, scattering, snr, resolution
         """
 
         # Get the parent directory of the input directory
         parentDirectory = os.path.dirname(self.directoryPath)
 
+        # Create the results directory if it does not exist
+        resultsDirectory = os.path.join(parentDirectory, 'results')
+        os.makedirs(resultsDirectory, exist_ok=True)
+
+        # Generate the result file name
+        resultFileName = f"{self.sampleMaterial}_{int(self.sampleDiameterMm)}mm_{int(self.kvp)}kv_bin{int(self.binning)}_result.json"
+        resultFilePath = os.path.join(resultsDirectory, resultFileName)
+
+        # Check if the result file already exists
+        if os.path.exists(resultFilePath):
+            print(f"Result file already exists: {resultFilePath}")
+            return
+
         # Generate the log file name
-        logFileName = f"{self.sampleMaterial}_{int(self.sampleDiameterMm)}mm_{int(self.kvp)}kv_log.txt"
-        logFilePath = os.path.join(parentDirectory, logFileName)
+        logFileName = f"{self.sampleMaterial}_{int(self.sampleDiameterMm)}mm_{int(self.kvp)}kv_bin{int(self.binning)}_log.txt"
+        logFilePath = os.path.join(resultsDirectory, logFileName)
 
         # Open the log file
         with open(logFilePath, "w") as logFile:
@@ -154,11 +171,7 @@ class QualityMeasuresAnalyzer:
                     "Resolution": resolution
                 }
 
-                # Generate the output file name
-                resultFileName = f"{self.sampleMaterial}_{int(self.sampleDiameterMm)}mm_{int(self.kvp)}kv_result.json"
-                resultFilePath = os.path.join(parentDirectory, resultFileName)
-
-                # Save results to a file in the parent directory
+                # Save results to a file in the results directory
                 with open(resultFilePath, "w") as resultFile:
                     json.dump(results, resultFile, indent=4)
 
@@ -166,13 +179,39 @@ class QualityMeasuresAnalyzer:
 
 
 
-# if __name__ == '__main__':
-#     al33mm180kv = QualityMeasuresAnalyzer('/home/xilin/projects/recon_ws/EfficientScans/'
-#                                        'AL_33mm__180kV-80uA_bin2_450_CD1150mm', "Al", shape=(938, 938))
-#     al33mm180kv.analyseAll()
-
-
 if __name__ == '__main__':
-    pmma38mm60kv = QualityMeasuresAnalyzer('/home/xilin/projects/recon_ws/EfficientScans/'
-                                          'PMMA_38mm__60kV-150uA_bin2_450_CD1150mm', "pmma", shape=(938, 938))
-    pmma38mm60kv.analyseAll()
+
+    import os
+    import re
+
+    base_path = '/home/xilin/projects/recon_ws/EfficientScans/'
+
+    # Regular expression to match the folder name pattern
+    folder_pattern = re.compile(
+        r'(?P<sampleMaterial>[^_]+)_(?P<sampleDiameter>[^_]+)__(?P<kvp>[^-]+)-(?P<sourceCurrent>[^_]+)_bin(?P<binning>\d+)_.*')
+
+    for folder_name in os.listdir(base_path):
+        folder_path = os.path.join(base_path, folder_name)
+        if os.path.isdir(folder_path):
+            match = folder_pattern.match(folder_name)
+            if match:
+                sampleMaterial = match.group('sampleMaterial')
+                binning = int(match.group('binning'))
+                # Corrected shape assignment based on binning value
+                if binning == 1:
+                    shape = (1876, 1876)
+                elif binning == 2:
+                    shape = (938, 938)
+                else:
+                    raise ValueError(f"Unsupported binning value: {binning}")
+
+                print(f"\nProcessing folder: {folder_name}")
+                print(f"Sample Material: {sampleMaterial}, Binning: {binning}, Shape: {shape}")
+
+                analyzer = QualityMeasuresAnalyzer(folder_path, sampleMaterial, shape)
+                analyzer.analyseAll()
+            else:
+                print(f"\nFolder name does not match pattern: {folder_name}")
+
+
+
